refactoring into validator
